+++ enclone_main/src/lib.rs
@@ -1,5 +1,6 @@
 // Copyright (c) 2021 10x Genomics, Inc. All rights reserved.
 
+pub mod doublets;
 pub mod inconsistent;
 pub mod main_enclone;
 pub mod setup;
