+++ enclone_main/tests/enclone_test.rs
@@ -2200,7 +2200,7 @@ fn test_cpu_usage() {
 #[test]
 fn test_source_code_file_length() {
     PrettyTrace::new().on();
-    const MAX_RS_LINES: usize = 2421;
+    const MAX_RS_LINES: usize = 2235;
     let top = dir_list("..");
     let mut dirs = Vec::<String>::new();
     for d in top.iter() {
